separate db snapshots logic
