+++ services/web/server/src/simcore_service_webserver/meta_api_handlers_snapshots.py
@@ -15,7 +15,7 @@ from .meta_api_handlers_base import (
     handle_request_errors,
 )
 from .meta_core_snapshots import ProjectDict, take_snapshot
-from .meta_db import ProjectsRepository, SnapshotsRepository
+from .meta_db_snapshots import ProjectsRepository, SnapshotsRepository
 from .meta_models_snapshots import Snapshot, SnapshotPatchBody, SnapshotResource
 from .projects import projects_api
 from .security_decorators import permission_required
